test 3dbool : find a problem ,no matter where, can't accomplish in loop clear iterator. In there is planar during Interation generator new exisplanars and delete outdate planars .
So today write a Data::structcopy() . when in meet() and after load() work. Take the function to copy every struct and maintain their relation such as inPlanar , points endpoints and so on. after complie went back last error point. next time change PastPoint and PastEdge. so that change Planar and needn't delete old and create new so that break existplanar loop.